@@ -44,18 +44,15 @@
     "react": "^17.0.1",
     "react-css-modules": "^4.7.11",
     "react-dom": "^17.0.1",
-    "react-hot-loader": "^4.13.0",
-    "react-i18next": "^11.11.4",
-    "react-router-cache-route": "^1.12.1",
-    "react-router-dom": "^5.2.0",
+    "react-i18next": "^11.16.2",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-spring": "^9.4.4",
-    "recoil": "^0.5.0",
+    "recoil": "^0.6.1",
     "stream-browserify": "^3.0.0",
-    "sweetalert2": "^11.1.9",
-    "sweetalert2-react-content": "^4.1.1",
-    "swr": "^0.5.6",
-    "ttag": "1.7.15",
-    "typescript": "^4.3.5",
+    "sweetalert2": "^11.4.8",
+    "sweetalert2-react-content": "^4.2.0",
+    "typescript": "^4.6.3",
     "use-immer": "^0.6.0",
     "webpack": "^5.11.0",
     "webpack-merge": "^5.7.3"
@@ -66,47 +63,41 @@
     "@babel/preset-react": "^7.12.10",
     "@babel/preset-typescript": "^7.12.7",
     "@types/jest": "^26.0.20",
-    "@types/js-cookie": "^2.2.7",
     "@types/lodash-es": "^4.17.6",
     "@types/node": "^16.7.1",
-    "@types/react": "^17.0.17",
+    "@types/react": "^17.0.43",
     "@types/react-css-modules": "^4.6.4",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router": "^5.1.16",
-    "@types/react-router-config": "^5.0.1",
-    "@types/react-router-dom": "^5.1.8",
-    "@typescript-eslint/eslint-plugin": "^4.9.1",
-    "@typescript-eslint/parser": "^4.1.1",
-    "autoprefixer": "^10.3.2",
+    "@types/react-dom": "^17.0.14",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
+    "autoprefixer": "^10.4.4",
     "babel-plugin-import": "^1.13.3",
-    "babel-plugin-ttag": "^1.7.26",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^9.0.1",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^10.2.4",
     "css-loader": "^5.0.1",
     "cssnano": "^5.0.8",
-    "eslint": "^8.11.0",
+    "eslint": "^8.12.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-webpack-plugin": "^3.1.1",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^4.5.0",
+    "html-webpack-plugin": "^5.5.0",
     "less": "^4.1.1",
     "less-loader": "^10.0.1",
     "lint-staged": "^12.3.5",
-    "mini-css-extract-plugin": "^2.2.0",
+    "mini-css-extract-plugin": "^2.6.0",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-loader": "^6.1.1",
+    "postcss-loader": "^6.2.1",
     "postcss-normalize": "^10.0.0",
     "postcss-preset-env": "^7.4.3",
-    "prettier": "^2.3.2",
-    "react-router-config": "^5.1.1",
-    "style-loader": "^2.0.0",
+    "prettier": "^2.6.1",
+    "style-loader": "^3.3.1",
     "ts-jest": "^26.4.4",
     "url-loader": "^4.1.1",
     "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.3.0",
-    "webpack-dev-server": "^3.11.0"
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   },
   "lint-staged": {
     "*.js": "eslint --cache --fix"
