@@ -15,19 +15,26 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import { auth } from "@src/utils/auth";
-import React from "react";
-import { Redirect, Route } from "react-router";
+import { auth } from '@src/utils/auth';
+import React from 'react';
+import { Navigate, Route } from 'react-router';
 
 export function InitializedRoute({ children, ...rest }) {
-    const isPassportLogin = location.pathname.includes("/passport/login");
+    const isPassportLogin = location.pathname.includes('/passport/login');
     return (
-        <Route {...rest} render={props => {
-            if (!auth.checkInitialized()) {
-                return <Redirect to="/initialize" />;
-            } else {
-                return auth.checkLogin() || isPassportLogin ? children : <Redirect to="/passport/login" />;
-            }
-        }} />
-    )
+        <Route
+            {...rest}
+            render={() => {
+                if (!auth.checkInitialized()) {
+                    return <Navigate to="/initialize" />;
+                } else {
+                    return auth.checkLogin() || isPassportLogin ? (
+                        children
+                    ) : (
+                        <Route path="/" render={() => <Navigate to="/passport/login" />} />
+                    );
+                }
+            }}
+        />
+    );
 }
