@@ -15,8 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import React from 'react';
-import { useLocation, useHistory } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { Tabs } from 'antd';
 import { useTranslation } from 'react-i18next';
 
@@ -24,12 +23,12 @@ const { TabPane } = Tabs;
 
 export default function TabsHeader() {
     const { pathname } = useLocation();
-    const history = useHistory();
+    const navigate = useNavigate();
 
-    const {t} = useTranslation()
+    const { t } = useTranslation();
 
     const handleTabChange = (key: string) => {
-        history.replace(key);
+        navigate(key);
     };
 
     return (
