@@ -15,9 +15,9 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Card } from 'antd';
-import { Redirect, Route, Switch, useRouteMatch, useHistory } from 'react-router-dom';
+import { Navigate, Route, Routes, useNavigate } from 'react-router-dom';
 import styles from './settings.module.less';
 import { UserInfoContext } from '@src/common/common.context';
 import { Sidebar } from '@src/components/sidebar/sidebar';
@@ -27,15 +27,14 @@ import { User } from './user/user';
 import { useUserInfo } from '@src/hooks/use-userinfo.hooks';
 import LoadingLayout from './global/components/loading-layout';
 
-export function Settings(props: any) {
-    const match = useRouteMatch();
-    const history = useHistory();
+export function Settings() {
+    const navigate = useNavigate();
     const [loading, setLoading] = useState(true);
     const [userInfo] = useUserInfo();
     useEffect(() => {
         if (userInfo.id == null) return;
         if (userInfo.id != null && !userInfo.is_super_admin) {
-            history.push('/space');
+            navigate('/space');
             return;
         }
         setLoading(false);
@@ -49,10 +48,10 @@ export function Settings(props: any) {
                     <Card className={styles.card}>
                         <TabsHeader />
                         <LoadingLayout loading={loading} wrapperStyle={{ textAlign: 'center', marginTop: 200 }}>
-                            <Switch>
-                                <Route path={`${match.path}/user`} component={User} />
-                                <Redirect to={`${match.path}/user`} />
-                            </Switch>
+                            <Routes>
+                                <Route path="users/*" element={<User />} />
+                                <Route path="/" element={<Navigate replace to="users" />} />
+                            </Routes>
                         </LoadingLayout>
                     </Card>
                 </div>
