@@ -17,96 +17,71 @@
 
 /** @format */
 
-import React, { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect } from 'react';
 import styles from './table-content.module.less';
 import CSSModules from 'react-css-modules';
-import { Layout, Menu, Tabs, Button } from 'antd';
+import { Layout, Tabs } from 'antd';
 import { TableOutlined } from '@ant-design/icons';
 import { CommonHeader } from '@src/components/common-header/header';
 
-import { TableInfoResponse } from './table.interface';
-import { TableAPI } from './table.api';
-import { useHistory } from 'react-router-dom';
-
 import BaseInfo from './tabs/baseInfo';
 import { Schema } from './schema/schema';
-import DataImportTab from './tabs/data-import';
 import { TableInfoTabTypeEnum } from './table-content.data';
-import { Redirect, Route, Switch } from 'react-router-dom';
-import { pathToRegexp } from 'path-to-regexp';
+import { Navigate, Route, Routes, useMatch, useNavigate } from 'react-router-dom';
 import DataPreview from './tabs/data.pre';
-const { Content, Sider } = Layout;
+const { Content } = Layout;
 const iconTable = <TableOutlined />;
 import { isTableIdSame } from '@src/utils/utils';
 import { useTranslation } from 'react-i18next';
 const { TabPane } = Tabs;
-let id: any = '',
-    name: any = '';
+
 function TableContent(props: any) {
-    const { t } = useTranslation()
-    const history = useHistory();
-    const { match } = props;
+    const { t } = useTranslation();
+    console.log(props);
+    const navigate = useNavigate();
+    const [id, setId] = useState(() => localStorage.getItem('table_id'));
+    const [name, setName] = useState(() => localStorage.getItem('table_name'));
     const [dbId, setDbId] = useState<any>();
-    const matchedPath = pathToRegexp(`${match.path}/:tabType`).exec(props.location.pathname);
+    const match = useMatch('/meta/table/:tableId/:tabType');
     useEffect(() => {
-        id = localStorage.getItem('table_id');
-        name = localStorage.getItem('table_name');
         setDbId(localStorage.getItem('database_id'));
         isTableIdSame();
     }, [window.location.href]);
 
-    function refresh(router: any) {
+    function refresh() {
         //
     }
 
     function handleTabChange(activeTab: string) {
-        props.history.push({
-            pathname: `${match.url}/${activeTab}`,
+        navigate(`${activeTab}`, {
             state: { id: id, name: name },
         });
     }
 
     return (
         <Content styleName="table-main">
-            <CommonHeader title={name} icon={iconTable} callback={refresh}></CommonHeader>
+            <CommonHeader title={name as string} icon={iconTable} callback={refresh}></CommonHeader>
             <div styleName="table-content">
                 <Tabs
-                    // defaultActiveKey={`${ma}`}
-                    activeKey={matchedPath ? matchedPath[2] : TableInfoTabTypeEnum.BasicInfo}
+                    activeKey={match?.params.tabType ? match?.params.tabType : TableInfoTabTypeEnum.BasicInfo}
                     onChange={(activeTab: string) => handleTabChange(activeTab)}
                 >
-                    <TabPane tab={t`BasicInfo`} key={TableInfoTabTypeEnum.BasicInfo}>
-                        {/* <BaseInfo tableInfo={tableInfo} /> */}
-                    </TabPane>
-                    <TabPane tab={t`DataPreview`} key={TableInfoTabTypeEnum.DataPreview}>
-                        {/* 12312 */}
-                    </TabPane>
+                    <TabPane tab={t`BasicInfo`} key={TableInfoTabTypeEnum.BasicInfo}></TabPane>
+                    <TabPane tab={t`DataPreview`} key={TableInfoTabTypeEnum.DataPreview}></TabPane>
                     <TabPane tab="Schema" key={TableInfoTabTypeEnum.Schema}></TabPane>
                 </Tabs>
-                <Switch>
+                <Routes>
                     <Route
-                        path={`${match.path}/${TableInfoTabTypeEnum.BasicInfo}`}
-                        render={props => <BaseInfo tableId={match.params.tableId} />}
-                    />
-                    <Route
-                        path={`${match.path}/${TableInfoTabTypeEnum.DataPreview}`}
-                        render={props => <DataPreview tableId={match.params.tableId} />}
+                        path={TableInfoTabTypeEnum.BasicInfo}
+                        element={<BaseInfo tableId={match?.params.tableId} />}
                     />
                     <Route
-                        path={`${match.path}/${TableInfoTabTypeEnum.DataImport}`}
-                        render={props => <DataImportTab tableId={match.params.tableId} tableName={name} />}
+                        path={TableInfoTabTypeEnum.DataPreview}
+                        element={<DataPreview tableId={match?.params.tableId} />}
                     />
-                    <Route
-                        path={`${match.path}/${TableInfoTabTypeEnum.Schema}`}
-                        render={props => <Schema tableId={match.params.tableId} />}
-                    />
-                    <Redirect
-                        to={{
-                            pathname: `${match.path}/${TableInfoTabTypeEnum.BasicInfo}`,
-                            state: { id: id, name: name },
-                        }}
-                    />
-                </Switch>
+                    <Route path={TableInfoTabTypeEnum.Schema} element={<Schema tableId={match?.params.tableId} />} />
+                    <Route path="/" element={<Navigate replace to={TableInfoTabTypeEnum.BasicInfo} />} />
+                </Routes>
             </div>
         </Content>
     );
