@@ -15,40 +15,27 @@
 // specific language governing permissions and limitations
 // under the License.
 
-const webpackBaseConfig = require("./webpack.base.config");
-const { merge } = require("webpack-merge");
-const SERVER_RD = 'http://127.0.0.1:8880';
+const webpackBaseConfig = require('./webpack.base.config');
+const { merge } = require('webpack-merge');
+const SERVER_RD = 'http://st01-meg-lijiu02-414.st01.baidu.com:8880/';
 
 module.exports = merge(webpackBaseConfig, {
-  mode: "development",
-  devtool: "eval-source-map",
-  // 开发服务配置
-  devServer: {
-    host: "localhost",
-    port: 8084,
-    publicPath: "/",
-    disableHostCheck: true,
-    useLocalIp: false,
-    open: true,
-    overlay: true,
-    hot: true,
-    stats: {
-      assets: false,
-      colors: true,
-      modules: false,
-      children: false,
-      chunks: false,
-      chunkModules: false,
-      entrypoints: false,
+    mode: 'development',
+    devtool: 'eval-source-map',
+    // 开发服务配置
+    devServer: {
+        host: 'localhost',
+        port: 8084,
+        open: true,
+        hot: true,
+        proxy: {
+            '/api': {
+                target: SERVER_RD,
+                changeOrigin: true,
+                secure: false,
+                // pathRewrite: {'^/api': '/api'}
+            },
+        },
+        historyApiFallback: true,
     },
-    proxy: {
-      '/api': {
-          target: SERVER_RD,
-          changeOrigin: true,
-          secure: false,
-          // pathRewrite: {'^/api': '/api'}
-      },
-    },
-    historyApiFallback: true,
-  },
 });
