@@ -16,13 +16,13 @@
 // under the License.
 
 import { getDefaultPageSize } from './utils/utils';
-import version from '../version.json';
+import packageInfo from '../package.json';
 
 export const TABLE_DELAY = 150;
 export const PAGESIZE_OPTIONS = ['10', '20', '50', '100', '200'];
 export const PAGINATION = {
     current: 1,
     pageSize: getDefaultPageSize(),
-    // total: 0
+    total: 0,
 };
-export const VERSION = `v${version.version}`;
+export const VERSION = `v${packageInfo.version}`;
